+++ R/reporter-llm.R
@@ -1,8 +1,13 @@
 #' Report test progress for LLMs
 #'
+#' @description
 #' `LlmReporter` is designed for use with Large Language Models (LLMs).
 #' It reports problems (warnings, skips, errors, and failures) as they
 #' occur and the total number of successes at the end.
+#'
+#' `LlmReporter` is used by default when tests are run by a coding agent.
+#' Currently we detect Claude Code, Cursor, and Gemini CLI.
+#' If using another tool, configure it to set env var `AGENT=1`.
 #'
 #' @export
 #' @family reporters
@@ -17,6 +22,8 @@ LlmReporter <- R6::R6Class(
         ...
       )
       self$width <- 20
+      self$rstudio <- FALSE
+      self$hyperlinks <- FALSE
     },
 
     # No header
@@ -79,3 +86,10 @@ LlmReporter <- R6::R6Class(
     }
   )
 )
+
+is_llm <- function() {
+  nzchar(Sys.getenv("AGENT")) ||
+    nzchar(Sys.getenv("CLAUDECODE")) ||
+    nzchar(Sys.getenv("GEMINI_CLI")) ||
+    nzchar(Sys.getenv("CURSOR_AGENT"))
+}